fixed insert/relationship issue, changed yelp datatype

+++ db/migrate/20180925205751_create_third_party_ratings.rb
@@ -1,7 +1,7 @@
 class CreateThirdPartyRatings < ActiveRecord::Migration[5.2]
   def change
     create_table :third_party_ratings do |t|
-      t.integer :yelp
+      t.decimal :yelp
 
       t.timestamps
     end
